pyproject.toml
[project]
name = "tidymon"
version = "0.2.0"
description = "Windows 디지털 환경 정리 모니터 (폴더 감시 + 북마크 정리)"
requires-python = ">=3.11"
dependencies = [
    "winotify",
    "pyyaml",
    "pystray",
    "Pillow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
tidymon = "monitor:run"
tidymon-scheduler = "setup_scheduler:register"
tidymon-tray = "tray:main"
